requirements.txt
albumentations
black
easydict
flake8
ipykernel
matplotlib
ndjson
numpy
opencv-python
openpyxl
Pillow
pycocotools
PyYAML
scikit-learn
scikit-image
seaborn
segmentation-models-pytorch
tensorboardX
timm
tqdm
urllib3
yacs